rebase

+++ frappe/desk/doctype/desktop_icon/desktop_icon.py
@@ -9,6 +9,8 @@ from frappe import _
 import json
 import random
 from frappe.model.document import Document
+from six import iteritems, string_types
+from frappe.utils.user import UserPermissions
 
 class DesktopIcon(Document):
 	def validate(self):
@@ -30,22 +32,34 @@ def get_desktop_icons(user=None):
 
 	if not user_icons:
 		fields = ['module_name', 'hidden', 'label', 'link', 'type', 'icon', 'color',
-			'_doctype', 'idx', 'force_show', 'reverse', 'custom', 'standard', 'blocked']
+			'_doctype', '_report', 'idx', 'force_show', 'reverse', 'custom', 'standard', 'blocked']
+
+		active_domains = frappe.get_active_domains()
+
+		blocked_doctypes = frappe.get_all("DocType", filters={
+			"ifnull(restrict_to_domain, '')": ("not in", ",".join(active_domains))
+		}, fields=["name"])
+
+		blocked_doctypes = [ d.get("name") for d in blocked_doctypes ]
 
 		standard_icons = frappe.db.get_all('Desktop Icon',
 			fields=fields, filters={'standard': 1})
 
 		standard_map = {}
 		for icon in standard_icons:
+			if icon._doctype in blocked_doctypes:
+				icon.blocked = 1
 			standard_map[icon.module_name] = icon
 
 		user_icons = frappe.db.get_all('Desktop Icon', fields=fields,
 			filters={'standard': 0, 'owner': user})
 
-
 		# update hidden property
 		for icon in user_icons:
 			standard_icon = standard_map.get(icon.module_name, None)
+
+			if icon._doctype in blocked_doctypes:
+				icon.blocked = 1
 
 			# override properties from standard icon
 			if standard_icon:
@@ -53,10 +67,11 @@ def get_desktop_icons(user=None):
 					if standard_icon.get(key):
 						icon[key] = standard_icon.get(key)
 
+
 				if standard_icon.blocked:
 					icon.hidden = 1
 
-					# flag for modules_setup page
+					# flag for modules_select dialog
 					icon.hidden_in_standard = 1
 
 				elif standard_icon.force_show:
@@ -80,7 +95,7 @@ def get_desktop_icons(user=None):
 				icon.hidden = 1
 
 		# sort by idx
-		user_icons.sort(lambda a, b: 1 if a.idx > b.idx else -1)
+		user_icons.sort(key = lambda a: a.idx)
 
 		# translate
 		for d in user_icons:
@@ -91,10 +106,10 @@ def get_desktop_icons(user=None):
 	return user_icons
 
 @frappe.whitelist()
-def add_user_icon(_doctype, label=None, link=None, type='link', standard=0):
+def add_user_icon(_doctype, _report=None, label=None, link=None, type='link', standard=0):
 	'''Add a new user desktop icon to the desktop'''
 
-	if not label: label = frappe._(_doctype)
+	if not label: label = _doctype or _report
 	if not link: link = 'List/{0}'.format(_doctype)
 
 	# find if a standard icon exists
@@ -112,8 +127,13 @@ def add_user_icon(_doctype, label=None, link=None, type='link', standard=0):
 			frappe.session.user)[0][0] or \
 			frappe.db.sql('select count(*) from `tabDesktop Icon` where standard=1')[0][0]
 
-		module = frappe.db.get_value('DocType', _doctype, 'module')
-		module_icon = frappe.get_value('Desktop Icon', {'standard':1, 'module_name':module},
+		if not frappe.db.get_value("Report", _report):
+			_report = None
+			userdefined_icon = frappe.db.get_value('DocType', _doctype, ['icon','color','module'], as_dict=True)
+		else:
+			userdefined_icon = frappe.db.get_value('Report', _report, ['icon','color','module'], as_dict=True)
+
+		module_icon = frappe.get_value('Desktop Icon', {'standard':1, 'module_name':userdefined_icon.module},
 			['name', 'icon', 'color', 'reverse'], as_dict=True)
 
 		if not module_icon:
@@ -130,8 +150,9 @@ def add_user_icon(_doctype, label=None, link=None, type='link', standard=0):
 				'link': link,
 				'type': type,
 				'_doctype': _doctype,
-				'icon': module_icon.icon,
-				'color': module_icon.color,
+				'_report': _report,
+				'icon': userdefined_icon.icon or module_icon.icon,
+				'color': userdefined_icon.color or module_icon.color,
 				'reverse': module_icon.reverse,
 				'idx': idx + 1,
 				'custom': 1,
@@ -141,7 +162,9 @@ def add_user_icon(_doctype, label=None, link=None, type='link', standard=0):
 
 			icon_name = new_icon.name
 
-		except Exception, e:
+		except frappe.UniqueValidationError as e:
+			frappe.throw(_('Desktop Icon already exists'))
+		except Exception as e:
 			raise e
 
 	return icon_name
@@ -150,7 +173,7 @@ def add_user_icon(_doctype, label=None, link=None, type='link', standard=0):
 @frappe.whitelist()
 def set_order(new_order, user=None):
 	'''set new order by duplicating user icons (if user is set) or set global order'''
-	if isinstance(new_order, basestring):
+	if isinstance(new_order, string_types):
 		new_order = json.loads(new_order)
 	for i, module_name in enumerate(new_order):
 		if module_name not in ('Explore',):
@@ -175,11 +198,13 @@ def set_desktop_icons(visible_list, ignore_duplicate=True):
 	if the desktop icon does not exist and the name is a DocType, then will create
 	an icon for the doctype'''
 
-	# clear all custom
-	frappe.db.sql('delete from `tabDesktop Icon` where standard=0')
+	# clear all custom only if setup is not complete
+	if not int(frappe.defaults.get_defaults().setup_complete or 0):
+		frappe.db.sql('delete from `tabDesktop Icon` where standard=0')
 
-	# set all as blocked
-	frappe.db.sql('update `tabDesktop Icon` set blocked=0, hidden=1')
+	# set standard as blocked and hidden if setting first active domain
+	if not frappe.flags.keep_desktop_icons:
+		frappe.db.sql('update `tabDesktop Icon` set blocked=0, hidden=1 where standard=1')
 
 	# set as visible if present, or add icon
 	for module_name in visible_list:
@@ -190,7 +215,7 @@ def set_desktop_icons(visible_list, ignore_duplicate=True):
 			if frappe.db.exists('DocType', module_name):
 				try:
 					add_user_icon(module_name, standard=1)
-				except frappe.UniqueValidationError, e:
+				except frappe.UniqueValidationError as e:
 					if not ignore_duplicate:
 						raise e
 					else:
@@ -207,7 +232,7 @@ def set_hidden_list(hidden_list, user=None):
 	'''Sets property `hidden`=1 in **Desktop Icon** for given user.
 	If user is None then it will set global values.
 	It will also set the rest of the icons as shown (`hidden` = 0)'''
-	if isinstance(hidden_list, basestring):
+	if isinstance(hidden_list, string_types):
 		hidden_list = json.loads(hidden_list)
 
 	# set as hidden
@@ -271,7 +296,7 @@ def make_user_copy(module_name, user):
 	standard_name = frappe.db.get_value('Desktop Icon', {'module_name': module_name, 'standard': 1})
 
 	if not standard_name:
-		frappe.throw('{0} not found'.format(module_name), frappe.DoesNotExistError)
+		frappe.throw(_('{0} not found').format(module_name), frappe.DoesNotExistError)
 
 	original = frappe.get_doc('Desktop Icon', standard_name)
 
@@ -304,7 +329,7 @@ def sync_from_app(app):
 
 	if isinstance(modules, dict):
 		modules_list = []
-		for m, desktop_icon in modules.iteritems():
+		for m, desktop_icon in iteritems(modules):
 			desktop_icon['module_name'] = m
 			modules_list.append(desktop_icon)
 	else:
@@ -329,9 +354,75 @@ def sync_from_app(app):
 			m['_doctype'] = m.pop('doctype')
 
 		desktop_icon.update(m)
-		desktop_icon.save()
+		try:
+			desktop_icon.save()
+		except frappe.exceptions.UniqueValidationError:
+			pass
 
 	return modules_list
+
+@frappe.whitelist()
+def update_icons(hidden_list, user=None):
+	"""update modules"""
+	if not user:
+		frappe.only_for('System Manager')
+
+	set_hidden_list(hidden_list, user)
+	frappe.msgprint(frappe._('Updated'), indicator='green', title=_('Success'), alert=True)
+
+def get_context(context):
+	context.icons = get_user_icons(frappe.session.user)
+	context.user = frappe.session.user
+
+	if 'System Manager' in frappe.get_roles():
+		context.users = frappe.db.get_all('User', filters={'user_type': 'System User', 'enabled': 1},
+			fields = ['name', 'first_name', 'last_name'])
+
+@frappe.whitelist()
+def get_module_icons(user=None):
+	if user != frappe.session.user:
+		frappe.only_for('System Manager')
+
+	if not user:
+		icons = frappe.db.get_all('Desktop Icon',
+			fields='*', filters={'standard': 1}, order_by='idx')
+	else:
+		frappe.cache().hdel('desktop_icons', user)
+		icons = get_user_icons(user)
+
+	for icon in icons:
+		icon.value = frappe.db.escape(_(icon.label or icon.module_name))
+
+	return {'icons': icons, 'user': user}
+
+def get_user_icons(user):
+	'''Get user icons for module setup page'''
+	user_perms = UserPermissions(user)
+	user_perms.build_permissions()
+
+	from frappe.boot import get_allowed_pages
+	allowed_pages = get_allowed_pages()
+
+	icons = []
+	for icon in get_desktop_icons(user):
+		add = True
+		if icon.hidden_in_standard:
+			add = False
+
+		if not icon.custom:
+			if icon.module_name=='Learn':
+				pass
+
+			elif icon.type=="page" and icon.link not in allowed_pages:
+				add = False
+
+			elif icon.type=="module" and icon.module_name not in user_perms.allow_modules:
+				add = False
+
+		if add:
+			icons.append(icon)
+
+	return icons
 
 palette = (
 	('#FFC4C4',),
@@ -383,3 +474,16 @@ palette = (
 	('#4F8EA8', 1),
 	('#428B46', 1)
 )
+
+@frappe.whitelist()
+def hide(name, user = None):
+	if not user:
+		user = frappe.session.user
+
+	try:
+		set_hidden(name, user, hidden = 1)
+		clear_desktop_icons_cache()
+	except Exception:
+		return False
+
+	return True
