+++ frappe/utils/install.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2015, Frappe Technologies Pvt. Ltd. and Contributors
 # MIT License. See license.txt
 
-from __future__ import unicode_literals
+from __future__ import unicode_literals, print_function
 
 import frappe
 import getpass
@@ -39,6 +39,10 @@ def after_install():
 	# setup wizard now in frappe
 	frappe.db.set_default('desktop:home_page', 'setup-wizard')
 
+	# clear test log
+	with open(frappe.get_site_path('.test_log'), 'w') as f:
+		f.write('')
+
 	frappe.db.commit()
 
 def install_basic_docs():
@@ -46,22 +50,23 @@ def install_basic_docs():
 	install_docs = [
 		{'doctype':'User', 'name':'Administrator', 'first_name':'Administrator',
 			'email':'admin@example.com', 'enabled':1, "is_admin": 1,
-			'user_roles': [{'role': 'Administrator'}]
+			'roles': [{'role': 'Administrator'}]
 		},
 		{'doctype':'User', 'name':'Guest', 'first_name':'Guest',
 			'email':'guest@example.com', 'enabled':1, "is_guest": 1,
-			'user_roles': [{'role': 'Guest'}]
+			'roles': [{'role': 'Guest'}]
 		},
 		{'doctype': "Role", "role_name": "Report Manager"},
+		{'doctype': "Role", "role_name": "Translator"},
 		{'doctype': "Workflow State", "workflow_state_name": "Pending",
 			"icon": "question-sign", "style": ""},
 		{'doctype': "Workflow State", "workflow_state_name": "Approved",
 			"icon": "ok-sign", "style": "Success"},
 		{'doctype': "Workflow State", "workflow_state_name": "Rejected",
 			"icon": "remove", "style": "Danger"},
-		{'doctype': "Workflow Action", "workflow_action_name": "Approve"},
-		{'doctype': "Workflow Action", "workflow_action_name": "Reject"},
-		{'doctype': "Workflow Action", "workflow_action_name": "Review"},
+		{'doctype': "Workflow Action Master", "workflow_action_name": "Approve"},
+		{'doctype': "Workflow Action Master", "workflow_action_name": "Reject"},
+		{'doctype': "Workflow Action Master", "workflow_action_name": "Review"},
 		{'doctype': "Email Domain", "domain_name":"example.com", "email_id": "account@example.com", "password": "pass", "email_server": "imap.example.com","use_imap": 1, "smtp_server": "smtp.example.com"},
 		{'doctype': "Email Account", "domain":"example.com", "email_id": "notifications@example.com", "default_outgoing": 1},
 		{'doctype': "Email Account", "domain":"example.com", "email_id": "replies@example.com", "default_incoming": 1}
@@ -78,7 +83,7 @@ def get_admin_password():
 		admin_password = getpass.getpass("Set Administrator password: ")
 		admin_password2 = getpass.getpass("Re-enter Administrator password: ")
 		if not admin_password == admin_password2:
-			print "\nPasswords do not match"
+			print("\nPasswords do not match")
 			return ask_admin_password()
 		return admin_password
 
@@ -89,8 +94,28 @@ def get_admin_password():
 
 
 def before_tests():
+	if len(frappe.get_installed_apps()) > 1:
+		# don't run before tests if any other app is installed
+		return
+
 	frappe.db.sql("delete from `tabCustom Field`")
 	frappe.db.sql("delete from `tabEvent`")
+	frappe.db.commit()
+	frappe.clear_cache()
+
+	# complete setup if missing
+	from frappe.desk.page.setup_wizard.setup_wizard import setup_complete
+	if not int(frappe.db.get_single_value('System Settings', 'setup_complete') or 0):
+		setup_complete({
+			"language"			:"english",
+			"email"				:"test@erpnext.com",
+			"full_name"			:"Test User",
+			"password"			:"test",
+			"country"			:"United States",
+			"timezone"			:"America/New_York",
+			"currency"			:"USD"
+		})
+
 	frappe.db.commit()
 	frappe.clear_cache()
 
@@ -105,7 +130,7 @@ def import_country_and_currency():
 		country = frappe._dict(data[name])
 		add_country_and_currency(name, country)
 
-	print
+	print("")
 
 	# enable frequently used currencies
 	for currency in ("INR", "USD", "GBP", "EUR", "AED", "AUD", "JPY", "CNY", "CHF"):
@@ -133,4 +158,3 @@ def add_country_and_currency(name, country):
 			"number_format": country.number_format,
 			"docstatus": 0
 		}).db_insert()
-
